added perm check to ?say

+++ commands/misc/say.js
@@ -1,4 +1,5 @@
 module.exports.run =(client, message, args) => {
+   if(message.member.hasPermission("MANAGE_MESSAGES")) return
    message.channel.send(args.join(" "));
    message.delete();
 };
